Move JumpToLastMessageButton to separate file; Move isLastReceivedMessage to utils; Move all CSS for button into the styles file;
