+++ src/script/util/conversationMessages.ts
@@ -21,6 +21,8 @@ import {Asset} from 'src/script/entity/message/Asset';
 import type {FileAsset as FileAssetType} from 'src/script/entity/message/FileAsset';
 
 import {AssetType} from '../assets/AssetType';
+import {Conversation} from '../entity/Conversation';
+import type {Message} from '../entity/message/Message';
 
 interface MessageDataType {
   senderName: string;
@@ -55,3 +57,7 @@ export function getMessageAriaLabel({senderName, displayTimestampShort, assets}:
     }
   });
 }
+
+export const isLastReceivedMessage = (messageEntity: Message, conversationEntity: Conversation): boolean => {
+  return messageEntity.timestamp() >= conversationEntity.last_event_timestamp();
+};
